Create file drain.go
